Removed AbstractBlockSummary (redundant) and IPowListener (unused)
