+++ modAionImpl/src/org/aion/zero/impl/types/AionBlockSummary.java
@@ -4,22 +4,30 @@ import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-
+import org.aion.log.AionLoggerFactory;
+import org.aion.log.LogEnum;
 import org.aion.mcf.blockchain.Block;
-import org.aion.mcf.types.AbstractBlockSummary;
 import org.aion.types.AionAddress;
 import org.aion.zero.impl.blockchain.BlockSummary;
 import org.aion.zero.types.AionTxExecSummary;
 import org.aion.zero.types.AionTxReceipt;
+import org.slf4j.Logger;
 
 /**
  * Modified to add transactions
  *
  * @author yao
  */
-public class AionBlockSummary
-        extends AbstractBlockSummary<AionTxReceipt, AionTxExecSummary>
-        implements BlockSummary {
+public class AionBlockSummary implements BlockSummary {
+
+    /* TODO: the getters for these parameters could be removed since they are final; requires removing the BlockSummary interface */
+    private final Block block;
+    private final Map<AionAddress, BigInteger> rewards;
+    private final List<AionTxReceipt> receipts;
+    private final List<AionTxExecSummary> summaries;
+    private BigInteger totalDifficulty = BigInteger.ZERO;
+
+    private final Logger LOG = AionLoggerFactory.getLogger(LogEnum.CONS.toString());
 
     public AionBlockSummary(
             Block block,
@@ -30,6 +38,38 @@ public class AionBlockSummary
         this.rewards = rewards;
         this.receipts = receipts;
         this.summaries = summaries;
+    }
+
+    public Block getBlock() {
+        return block;
+    }
+
+    public List<AionTxReceipt> getReceipts() {
+        return receipts;
+    }
+
+    public List<AionTxExecSummary> getSummaries() {
+        return summaries;
+    }
+
+    /**
+     * All the mining rewards paid out for this block, including the main block rewards, uncle
+     * rewards, and transaction fees.
+     */
+    public Map<AionAddress, BigInteger> getRewards() {
+        return rewards;
+    }
+
+    public void setTotalDifficulty(BigInteger totalDifficulty) {
+        this.totalDifficulty = totalDifficulty;
+
+        if (LOG.isTraceEnabled()) {
+            LOG.trace("The current total difficulty is: {}", totalDifficulty.toString());
+        }
+    }
+
+    public BigInteger getTotalDifficulty() {
+        return totalDifficulty;
     }
 
     public String toString() {
